"rp" alias... "re-prove"; re-run failed tests, in parallel.
Creates a nice workflow... "make test" to run all the unit-tests (and find a
bunch of failures).  Then, keep running "rp" each time you make a change, to
have it re-run *only* those tests that are still failing.  When you've got all
the tests fixed, it'll stop running. :)

A slight modification to this workflow would flush the saved state _before_
running the tests, and then re-running all of the tests at the very end as a
final sanity check before pushing:
    # remove the state file
    rm -f .prove

    # run tests, find breakage, fix them one by one
    make test
    rp          # (repeat)

    # re-run *all* the tests since ".prove" was removed
    st-prove --state=hot --jobs=4